+++ sdk/openai/azure-ai-openai/src/test/java/com/azure/ai/openai/NonAzureOpenAIAsyncClientTest.java
@@ -7,11 +7,18 @@ import com.azure.ai.openai.models.AudioTaskLabel;
 import com.azure.ai.openai.models.AudioTranscriptionFormat;
 import com.azure.ai.openai.models.AudioTranscriptionTimestampGranularity;
 import com.azure.ai.openai.models.AudioTranslationFormat;
+import com.azure.ai.openai.models.Batch;
+import com.azure.ai.openai.models.BatchCreateRequest;
+import com.azure.ai.openai.models.BatchStatus;
 import com.azure.ai.openai.models.ChatChoice;
 import com.azure.ai.openai.models.ChatCompletions;
 import com.azure.ai.openai.models.ChatCompletionsFunctionToolCall;
+import com.azure.ai.openai.models.ChatCompletionsFunctionToolSelection;
+import com.azure.ai.openai.models.ChatCompletionsNamedFunctionToolSelection;
 import com.azure.ai.openai.models.ChatCompletionsOptions;
 import com.azure.ai.openai.models.ChatCompletionsToolCall;
+import com.azure.ai.openai.models.ChatCompletionsToolSelection;
+import com.azure.ai.openai.models.ChatCompletionsToolSelectionPreset;
 import com.azure.ai.openai.models.ChatResponseMessage;
 import com.azure.ai.openai.models.ChatRole;
 import com.azure.ai.openai.models.Completions;
@@ -19,21 +26,29 @@ import com.azure.ai.openai.models.CompletionsFinishReason;
 import com.azure.ai.openai.models.CompletionsOptions;
 import com.azure.ai.openai.models.CompletionsUsage;
 import com.azure.ai.openai.models.Embeddings;
+import com.azure.ai.openai.models.FileDeletionStatus;
+import com.azure.ai.openai.models.FilePurpose;
+import com.azure.ai.openai.models.FileState;
 import com.azure.ai.openai.models.FunctionCall;
 import com.azure.ai.openai.models.FunctionCallConfig;
+import com.azure.ai.openai.models.OpenAIFile;
+import com.azure.ai.openai.models.PageableList;
 import com.azure.ai.openai.models.SpeechGenerationResponseFormat;
 import com.azure.core.credential.KeyCredential;
 import com.azure.core.exception.ClientAuthenticationException;
 import com.azure.core.exception.HttpResponseException;
 import com.azure.core.http.HttpClient;
 import com.azure.core.http.rest.RequestOptions;
+import com.azure.core.http.rest.Response;
 import com.azure.core.test.annotation.RecordWithoutRequestBody;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.CoreUtils;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -750,7 +765,9 @@ public class NonAzureOpenAIAsyncClientTest extends OpenAIClientTestBase {
     @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
     public void testGetChatCompletionsToolCall(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
         client = getNonAzureOpenAIAsyncClient(httpClient);
-        getChatWithToolCallRunnerForNonAzure((modelId, chatCompletionsOptions) -> StepVerifier.create(
+        getChatWithToolCallRunnerForNonAzure((modelId, chatCompletionsOptions) -> {
+            chatCompletionsOptions.setToolChoice(new ChatCompletionsToolSelection(ChatCompletionsToolSelectionPreset.AUTO));
+            StepVerifier.create(
                 client.getChatCompletionsWithResponse(modelId, chatCompletionsOptions, new RequestOptions())
                     .flatMap(response -> {
                         assertNotNull(response);
@@ -770,9 +787,9 @@ public class NonAzureOpenAIAsyncClientTest extends OpenAIClientTestBase {
                         ChatCompletionsFunctionToolCall functionToolCall = (ChatCompletionsFunctionToolCall) responseMessage.getToolCalls().get(0);
                         assertNotNull(functionToolCall);
                         assertFalse(functionToolCall.getFunction().getArguments() == null
-                                || functionToolCall.getFunction().getArguments().isEmpty());
+                            || functionToolCall.getFunction().getArguments().isEmpty());
                         return client.getChatCompletions(modelId, getChatCompletionsOptionWithToolCallFollowUp(
-                                functionToolCall, responseMessage.getContent()));
+                            functionToolCall, responseMessage.getContent()));
                     })).assertNext(followUpChatCompletions -> {
                         assertNotNull(followUpChatCompletions);
                         assertNotNull(followUpChatCompletions.getChoices());
@@ -783,62 +800,155 @@ public class NonAzureOpenAIAsyncClientTest extends OpenAIClientTestBase {
                         assertFalse(content == null || content.isEmpty());
                         assertEquals(followUpChatChoice.getMessage().getRole(), ChatRole.ASSISTANT);
                         assertEquals(followUpChatChoice.getFinishReason(), CompletionsFinishReason.STOPPED);
-                    }).verifyComplete());
+                    }).verifyComplete();
+        });
     }
 
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testGetChatCompletionToolCallChoiceExplicitToolName(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        getChatWithToolCallRunnerForNonAzure((modelId, chatCompletionsOptions) -> {
+            chatCompletionsOptions.setToolChoice(new ChatCompletionsToolSelection(
+                    new ChatCompletionsNamedFunctionToolSelection(
+                            new ChatCompletionsFunctionToolSelection("FutureTemperature"))));
+            StepVerifier.create(client.getChatCompletionsWithResponse(modelId, chatCompletionsOptions, new RequestOptions())).assertNext(response -> {
+                assertNotNull(response);
+                assertTrue(response.getStatusCode() >= 200 && response.getStatusCode() < 300);
+                ChatCompletions chatCompletions = response.getValue();
+                assertNotNull(chatCompletions);
+
+                assertTrue(chatCompletions.getChoices() != null && !chatCompletions.getChoices().isEmpty());
+                ChatChoice chatChoice = chatCompletions.getChoices().get(0);
+
+                assertNotNull(chatCompletions);
+                assertNotNull(chatCompletions.getChoices());
+                assertNotNull(chatChoice);
+                assertNotNull(chatChoice.getMessage());
+                ChatResponseMessage message = chatChoice.getMessage();
+                assertNull(message.getContent());
+                assertNotNull(message.getToolCalls().get(0));
+                assertInstanceOf(ChatCompletionsFunctionToolCall.class, message.getToolCalls().get(0));
+                ChatCompletionsFunctionToolCall functionToolCall = (ChatCompletionsFunctionToolCall) chatChoice.getMessage().getToolCalls().get(0);
+                assertEquals(functionToolCall.getFunction().getName(), "FutureTemperature");
+                assertTrue(functionToolCall.getFunction().getArguments().contains("Honolulu"));
+                assertEquals(chatChoice.getMessage().getRole(), ChatRole.ASSISTANT);
+                assertEquals(chatChoice.getFinishReason(), CompletionsFinishReason.STOPPED);
+            }).verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testGetChatCompletionToolCallChoiceNone(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        getChatWithToolCallRunnerForNonAzure((modelId, chatCompletionsOptions) -> {
+            chatCompletionsOptions.setToolChoice(new ChatCompletionsToolSelection(ChatCompletionsToolSelectionPreset.NONE));
+            StepVerifier.create(client.getChatCompletionsWithResponse(modelId, chatCompletionsOptions, new RequestOptions())).assertNext(response -> {
+                assertNotNull(response);
+                assertTrue(response.getStatusCode() >= 200 && response.getStatusCode() < 300);
+                ChatCompletions chatCompletions = response.getValue();
+                assertNotNull(chatCompletions);
+
+                assertTrue(chatCompletions.getChoices() != null && !chatCompletions.getChoices().isEmpty());
+                ChatChoice chatChoice = chatCompletions.getChoices().get(0);
+
+                assertNotNull(chatCompletions);
+                assertNotNull(chatCompletions.getChoices());
+                assertNotNull(chatChoice);
+                assertNotNull(chatChoice.getMessage());
+                ChatResponseMessage message = chatChoice.getMessage();
+                assertNotNull(message.getContent());
+                assertFalse(message.getContent().isEmpty());
+                assertNull(message.getToolCalls());
+                assertEquals(chatChoice.getMessage().getRole(), ChatRole.ASSISTANT);
+                assertEquals(chatChoice.getFinishReason(), CompletionsFinishReason.STOPPED);
+            }).verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testGetChatCompletionToolCallChoiceRequired(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        getChatWithToolCallRunnerForNonAzure((modelId, chatCompletionsOptions) -> {
+            chatCompletionsOptions.setToolChoice(new ChatCompletionsToolSelection(ChatCompletionsToolSelectionPreset.REQUIRED));
+            StepVerifier.create(client.getChatCompletionsWithResponse(modelId, chatCompletionsOptions, new RequestOptions())).assertNext(response -> {
+                assertNotNull(response);
+                assertTrue(response.getStatusCode() >= 200 && response.getStatusCode() < 300);
+                ChatCompletions chatCompletions = response.getValue();
+                assertNotNull(chatCompletions);
+
+                assertTrue(chatCompletions.getChoices() != null && !chatCompletions.getChoices().isEmpty());
+                ChatChoice chatChoice = chatCompletions.getChoices().get(0);
+
+                assertNotNull(chatCompletions);
+                assertNotNull(chatCompletions.getChoices());
+                assertNotNull(chatChoice);
+                assertNotNull(chatChoice.getMessage());
+                ChatResponseMessage message = chatChoice.getMessage();
+                assertNull(message.getContent());
+                assertNotNull(message.getToolCalls().get(0));
+                assertInstanceOf(ChatCompletionsFunctionToolCall.class, message.getToolCalls().get(0));
+                ChatCompletionsFunctionToolCall functionToolCall = (ChatCompletionsFunctionToolCall) chatChoice.getMessage().getToolCalls().get(0);
+                assertEquals(functionToolCall.getFunction().getName(), "FutureTemperature");
+                assertTrue(functionToolCall.getFunction().getArguments().contains("Honolulu"));
+                assertEquals(chatChoice.getMessage().getRole(), ChatRole.ASSISTANT);
+                assertEquals(chatChoice.getFinishReason(), CompletionsFinishReason.STOPPED);
+            }).verifyComplete();
+        });
+    }
     @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
     @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
     public void testGetChatCompletionsToolCallStreaming(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
         client = getNonAzureOpenAIAsyncClient(httpClient);
         getChatWithToolCallRunnerForNonAzure((modelId, chatCompletionsOptions) -> {
             StepVerifier.create(client.getChatCompletionsStream(modelId, chatCompletionsOptions)
-                    .collectList()
-                    .flatMapMany(chatCompletionsStream -> {
-                        StringBuilder argumentsBuilder = new StringBuilder();
-                        long totalStreamMessages = chatCompletionsStream.size();
-                        String functionName = null;
-                        String toolCallId = null;
-                        String content = null;
-                        assertTrue(totalStreamMessages > 0);
+                .collectList()
+                .flatMapMany(chatCompletionsStream -> {
+                    StringBuilder argumentsBuilder = new StringBuilder();
+                    long totalStreamMessages = chatCompletionsStream.size();
+                    String functionName = null;
+                    String toolCallId = null;
+                    String content = null;
+                    assertTrue(totalStreamMessages > 0);
 
-                        int i = 0;
-                        for (ChatCompletions chatCompletions : chatCompletionsStream) {
-                            List<ChatChoice> chatChoices = chatCompletions.getChoices();
-                            if (!chatChoices.isEmpty() && chatChoices.get(0) != null) {
-                                assertEquals(1, chatChoices.size());
-                                ChatChoice chatChoice = chatChoices.get(0);
-                                List<ChatCompletionsToolCall> toolCalls = chatChoice.getDelta().getToolCalls();
-                                if (toolCalls != null && !toolCalls.isEmpty()) {
-                                    assertEquals(1, toolCalls.size());
-                                    ChatCompletionsFunctionToolCall toolCall = (ChatCompletionsFunctionToolCall) toolCalls.get(0);
-                                    FunctionCall functionCall = toolCall.getFunction();
-                                    // TODO: It used to be first stream event but now second event, in NonAzure
-                                    // this data is only available in the first stream message, if at all
-                                    if (i == 1) {
-                                        content = chatChoice.getDelta().getContent();
-                                        functionName = functionCall.getName();
-                                        toolCallId = toolCall.getId();
-                                    }
-                                    argumentsBuilder.append(functionCall.getArguments());
+                    int i = 0;
+                    for (ChatCompletions chatCompletions : chatCompletionsStream) {
+                        List<ChatChoice> chatChoices = chatCompletions.getChoices();
+                        if (!chatChoices.isEmpty() && chatChoices.get(0) != null) {
+                            assertEquals(1, chatChoices.size());
+                            ChatChoice chatChoice = chatChoices.get(0);
+                            List<ChatCompletionsToolCall> toolCalls = chatChoice.getDelta().getToolCalls();
+                            if (toolCalls != null && !toolCalls.isEmpty()) {
+                                assertEquals(1, toolCalls.size());
+                                ChatCompletionsFunctionToolCall toolCall = (ChatCompletionsFunctionToolCall) toolCalls.get(0);
+                                FunctionCall functionCall = toolCall.getFunction();
+                                if (i == 0) {
+                                    content = chatChoice.getDelta().getContent();
+                                    functionName = functionCall.getName();
+                                    toolCallId = toolCall.getId();
                                 }
-                                if (i < totalStreamMessages - 1) {
-                                    assertNull(chatChoice.getFinishReason());
-                                } else {
-                                    assertEquals(CompletionsFinishReason.TOOL_CALLS, chatChoice.getFinishReason());
-                                }
+                                argumentsBuilder.append(functionCall.getArguments());
                             }
-                            i++;
+                            if (i < totalStreamMessages - 1) {
+                                assertNull(chatChoice.getFinishReason());
+                            } else {
+                                assertEquals(CompletionsFinishReason.TOOL_CALLS, chatChoice.getFinishReason());
+                            }
                         }
-                        assertFunctionToolCallArgs(argumentsBuilder.toString());
-                        FunctionCall functionCall = new FunctionCall(functionName, argumentsBuilder.toString());
-                        ChatCompletionsFunctionToolCall functionToolCall = new ChatCompletionsFunctionToolCall(toolCallId, functionCall);
+                        i++;
+                    }
+                    assertFunctionToolCallArgs(argumentsBuilder.toString());
+                    FunctionCall functionCall = new FunctionCall(functionName, argumentsBuilder.toString());
+                    ChatCompletionsFunctionToolCall functionToolCall = new ChatCompletionsFunctionToolCall(toolCallId, functionCall);
 
-                        ChatCompletionsOptions followUpChatCompletionsOptions = getChatCompletionsOptionWithToolCallFollowUp(
-                                functionToolCall, content);
+                    ChatCompletionsOptions followUpChatCompletionsOptions = getChatCompletionsOptionWithToolCallFollowUp(
+                            functionToolCall, content);
 
-                        return client.getChatCompletionsStream(modelId, followUpChatCompletionsOptions);
-                    })
-                    .collectList()
+                    return client.getChatCompletionsStream(modelId, followUpChatCompletionsOptions);
+                })
+                .collectList()
             ).assertNext(followupChatCompletionsStream -> {
                 StringBuilder contentBuilder = new StringBuilder();
                 long totalStreamFollowUpMessages = followupChatCompletionsStream.size();
@@ -869,12 +979,12 @@ public class NonAzureOpenAIAsyncClientTest extends OpenAIClientTestBase {
         client = getNonAzureOpenAIAsyncClient(httpClient);
         textToSpeechRunnerForNonAzure(((modelId, speechGenerationOptions) -> {
             StepVerifier.create(client.generateSpeechFromText(modelId, speechGenerationOptions))
-                    .assertNext(speech -> {
-                        assertNotNull(speech);
-                        byte[] bytes = speech.toBytes();
-                        assertNotNull(bytes);
-                        assertTrue(bytes.length > 0);
-                    }).verifyComplete();
+                .assertNext(speech -> {
+                    assertNotNull(speech);
+                    byte[] bytes = speech.toBytes();
+                    assertNotNull(bytes);
+                    assertTrue(bytes.length > 0);
+                }).verifyComplete();
         }));
     }
 
@@ -882,19 +992,18 @@ public class NonAzureOpenAIAsyncClientTest extends OpenAIClientTestBase {
     @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
     public void testTextToSpeechWithResponse(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
         client = getNonAzureOpenAIAsyncClient(httpClient);
-        textToSpeechRunnerForNonAzure(((modelId, speechGenerationOptions) -> {
+        textToSpeechRunnerForNonAzure(((modelId, speechGenerationOptions) ->
             StepVerifier.create(client.generateSpeechFromTextWithResponse(modelId,
-                            BinaryData.fromObject(speechGenerationOptions), new RequestOptions()))
-                    .assertNext(response -> {
-                        assertTrue(response.getStatusCode() > 0);
-                        assertNotNull(response.getHeaders());
-                        BinaryData speech = response.getValue();
-                        assertNotNull(speech);
-                        byte[] bytes = speech.toBytes();
-                        assertNotNull(bytes);
-                        assertTrue(bytes.length > 0);
-                    }).verifyComplete();
-        }));
+                BinaryData.fromObject(speechGenerationOptions), new RequestOptions()))
+            .assertNext(response -> {
+                assertTrue(response.getStatusCode() > 0);
+                assertNotNull(response.getHeaders());
+                BinaryData speech = response.getValue();
+                assertNotNull(speech);
+                byte[] bytes = speech.toBytes();
+                assertNotNull(bytes);
+                assertTrue(bytes.length > 0);
+            }).verifyComplete()));
     }
 
     @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
@@ -926,6 +1035,463 @@ public class NonAzureOpenAIAsyncClientTest extends OpenAIClientTestBase {
                         assertNotNull(bytes);
                         assertTrue(bytes.length > 0);
                     }).verifyComplete();
+        }));
+    }
+
+    // Files
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testTextFileOperations(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadTextFileRunner((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                    client.uploadFile(fileDetails, filePurpose)
+                        // Upload file
+                        .flatMap(uploadedFile -> {
+                            assertNotNull(uploadedFile);
+                            assertNotNull(uploadedFile.getId());
+                            return client.getFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Compare uploaded file with file from backend
+                        .flatMap(tuple -> {
+                            OpenAIFile fileFromBackend = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertNotNull(uploadedFile);
+                            assertNotNull(fileFromBackend);
+                            assertFileEquals(uploadedFile, fileFromBackend);
+                            return client.listFiles(FilePurpose.ASSISTANTS).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Check for existence of file when fetched by purpose
+                        .flatMap(tuple -> {
+                            List<OpenAIFile> files = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertTrue(files.stream().anyMatch(f -> f.getId().equals(uploadedFile.getId())));
+                            return client.deleteFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                        }))
+                // File deletion
+                .assertNext(tuple -> {
+                    FileDeletionStatus deletionStatus = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testImageFileOperations(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadImageFileRunner((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                    client.uploadFile(fileDetails, filePurpose)
+                        // Upload file
+                        .flatMap(uploadedFile -> {
+                            assertNotNull(uploadedFile);
+                            assertNotNull(uploadedFile.getId());
+                            return client.getFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Compare uploaded file with file from backend
+                        .flatMap(tuple -> {
+                            OpenAIFile fileFromBackend = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertNotNull(uploadedFile);
+                            assertNotNull(fileFromBackend);
+                            assertFileEquals(uploadedFile, fileFromBackend);
+                            return client.listFiles().zipWith(Mono.just(uploadedFile));
+                        })
+                        // Check for existence of files
+                        .flatMap(tuple -> {
+                            List<OpenAIFile> files = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertTrue(files.stream().anyMatch(f -> f.getId().equals(uploadedFile.getId())));
+                            return client.deleteFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                        }))
+                // File deletion
+                .assertNext(tuple -> {
+                    FileDeletionStatus deletionStatus = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testFineTuningJsonFileOperations(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadFineTuningJsonFileRunner((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                    client.uploadFile(fileDetails, filePurpose)
+                        // Upload file
+                        .flatMap(uploadedFile -> {
+                            assertNotNull(uploadedFile);
+                            assertNotNull(uploadedFile.getId());
+                            return client.getFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Compare uploaded file with file from backend
+                        .flatMap(tuple -> {
+                            OpenAIFile fileFromBackend = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertNotNull(uploadedFile);
+                            assertNotNull(fileFromBackend);
+                            assertFileEquals(uploadedFile, fileFromBackend);
+                            return client.listFiles(FilePurpose.FINE_TUNE).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Check for existence of file when fetched by purpose
+                        .flatMap(tuple -> {
+                            List<OpenAIFile> files = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertTrue(files.stream().anyMatch(f -> f.getId().equals(uploadedFile.getId())));
+                            return client.deleteFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                        }))
+                // File deletion
+                .assertNext(tuple -> {
+                    FileDeletionStatus deletionStatus = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testTextFileOperationsWithResponse(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadTextFileRunner((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                    client.uploadFile(fileDetails, filePurpose)
+                        // Upload file
+                        .flatMap(uploadedFile -> {
+                            assertNotNull(uploadedFile);
+                            assertNotNull(uploadedFile.getId());
+                            return client.getFileWithResponse(uploadedFile.getId(), new RequestOptions()).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Compare uploaded file with file from backend
+                        .flatMap(tuple -> {
+                            Response<OpenAIFile> response = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertNotNull(uploadedFile);
+                            assertNotNull(response);
+                            assertEquals(200, response.getStatusCode());
+                            OpenAIFile fileFromBackend = response.getValue();
+                            assertFileEquals(uploadedFile, fileFromBackend);
+
+                            RequestOptions requestOptions = new RequestOptions();
+                            requestOptions.addQueryParam("purpose", FilePurpose.ASSISTANTS.toString());
+                            return client.listFilesWithResponse(requestOptions).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Check for existence of file when fetched by purpose
+                        .flatMap(tuple -> {
+                            Response<List<OpenAIFile>> response = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertEquals(200, response.getStatusCode());
+                            List<OpenAIFile> files = response.getValue();
+                            assertTrue(files.stream().anyMatch(f -> f.getId().equals(uploadedFile.getId())));
+                            return client.deleteFileWithResponse(uploadedFile.getId(), new RequestOptions()).zipWith(Mono.just(uploadedFile));
+                        }))
+                // File deletion
+                .assertNext(tuple -> {
+                    Response<FileDeletionStatus> response = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+
+                    assertEquals(200, response.getStatusCode());
+                    FileDeletionStatus deletionStatus = response.getValue();
+
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testImageFileOperationsWithResponse(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadImageFileRunner((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                    client.uploadFile(fileDetails, filePurpose)
+                        // Upload file
+                        .flatMap(uploadedFile -> {
+                            assertNotNull(uploadedFile);
+                            assertNotNull(uploadedFile.getId());
+                            return client.getFileWithResponse(uploadedFile.getId(), new RequestOptions()).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Compare uploaded file with file from backend
+                        .flatMap(tuple -> {
+                            Response<OpenAIFile> response = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertNotNull(uploadedFile);
+                            assertNotNull(response);
+                            assertEquals(200, response.getStatusCode());
+                            OpenAIFile fileFromBackend = response.getValue();
+                            assertFileEquals(uploadedFile, fileFromBackend);
+
+                            RequestOptions requestOptions = new RequestOptions();
+                            requestOptions.addQueryParam("purpose", FilePurpose.ASSISTANTS.toString());
+                            return client.listFilesWithResponse(requestOptions).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Check for existence of file when fetched by purpose
+                        .flatMap(tuple -> {
+                            Response<List<OpenAIFile>> response = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertEquals(200, response.getStatusCode());
+                            List<OpenAIFile> files = response.getValue();
+                            assertTrue(files.stream().anyMatch(f -> f.getId().equals(uploadedFile.getId())));
+                            return client.deleteFileWithResponse(uploadedFile.getId(), new RequestOptions()).zipWith(Mono.just(uploadedFile));
+                        }))
+                // File deletion
+                .assertNext(tuple -> {
+                    Response<FileDeletionStatus> response = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+
+                    assertEquals(200, response.getStatusCode());
+                    FileDeletionStatus deletionStatus = response.getValue();
+
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testFineTuningJsonFileOperationsWithResponse(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadFineTuningJsonFileRunner((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                    client.uploadFile(fileDetails, filePurpose)
+                        // Upload file
+                        .flatMap(uploadedFile -> {
+                            assertNotNull(uploadedFile);
+                            assertNotNull(uploadedFile.getId());
+                            return client.getFileWithResponse(uploadedFile.getId(), new RequestOptions()).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Compare uploaded file with file from backend
+                        .flatMap(tuple -> {
+                            Response<OpenAIFile> response = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertNotNull(uploadedFile);
+                            assertNotNull(response);
+                            assertEquals(200, response.getStatusCode());
+                            OpenAIFile fileFromBackend = response.getValue();
+                            assertFileEquals(uploadedFile, fileFromBackend);
+
+                            RequestOptions requestOptions = new RequestOptions();
+                            requestOptions.addQueryParam("purpose", FilePurpose.FINE_TUNE.toString());
+                            return client.listFilesWithResponse(requestOptions).zipWith(Mono.just(uploadedFile));
+                        })
+                        // Check for existence of file when fetched by purpose
+                        .flatMap(tuple -> {
+                            Response<List<OpenAIFile>> response = tuple.getT1();
+                            OpenAIFile uploadedFile = tuple.getT2();
+
+                            assertEquals(200, response.getStatusCode());
+                            List<OpenAIFile> files = response.getValue();
+                            assertTrue(files.stream().anyMatch(f -> f.getId().equals(uploadedFile.getId())));
+                            return client.deleteFileWithResponse(uploadedFile.getId(), new RequestOptions()).zipWith(Mono.just(uploadedFile));
+                        }))
+                // File deletion
+                .assertNext(tuple -> {
+                    Response<FileDeletionStatus> response = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+
+                    assertEquals(200, response.getStatusCode());
+                    FileDeletionStatus deletionStatus = response.getValue();
+
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
+        });
+    }
+
+    // Batch
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testBatchOperations(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadBatchFileRunner(((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                client.uploadFile(fileDetails, filePurpose)
+                    // Upload file
+                    .flatMap(uploadedFile -> {
+                        assertNotNull(uploadedFile);
+                        assertNotNull(uploadedFile.getId());
+                        return client.getFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                    })
+                    // Compare uploaded file with file from backend
+                    .flatMap(tuple -> {
+                        OpenAIFile fileFromBackend = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(uploadedFile);
+                        assertNotNull(fileFromBackend);
+                        assertFileEquals(uploadedFile, fileFromBackend);
+                        return Mono.defer(() -> {
+                            if (fileFromBackend.getStatus() == FileState.PENDING) {
+                                return Mono.delay(Duration.ofSeconds(5))
+                                    .then(client.getFile(fileFromBackend.getId())).zipWith(Mono.just(uploadedFile));
+                            } else {
+                                return Mono.just(fileFromBackend).zipWith(Mono.just(uploadedFile));
+                            }
+                        });
+                    })
+                    // Create batch with file
+                    .flatMap(tuple -> {
+                        OpenAIFile fileFromBackend = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        return client.createBatch(new BatchCreateRequest("/v1/chat/completions", fileFromBackend.getId(), "24h"))
+                            .zipWith(Mono.just(uploadedFile));
+                    })
+                    // Looping getBatch until it's completed
+                    .flatMap(tuple -> {
+                        Batch batch = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(batch);
+                        assertNotNull(batch.getId());
+                        // TODO: make it to while loop
+                        return Mono.defer(() -> {
+                            if (batch.getStatus() == BatchStatus.VALIDATING
+                                || batch.getStatus() == BatchStatus.IN_PROGRESS
+                                || batch.getStatus() == BatchStatus.FINALIZING) {
+                                return Mono.delay(Duration.ofSeconds(20))
+                                    .then(client.getBatch(batch.getId())).zipWith(Mono.just(uploadedFile));
+                            } else {
+                                return Mono.just(batch).zipWith(Mono.just(uploadedFile));
+                            }
+                        });
+                    })
+                    // Get output file content
+                    .flatMap(tuple -> {
+                        Batch batch = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(batch);
+                        assertEquals(BatchStatus.COMPLETED, batch.getStatus());
+                        return client.getFileContent(batch.getOutputFileId()).zipWith(Mono.just(uploadedFile));
+                    })
+                    .flatMapMany(tuple -> {
+                        byte[] outputFile = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(outputFile);
+                        return client.listBatches().zipWith(Mono.just(uploadedFile));
+                    })
+                    // Delete uploaded file
+                    .flatMap(tuple -> {
+                        PageableList<Batch> batchPageableList = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(batchPageableList);
+                        assertFalse(CoreUtils.isNullOrEmpty(batchPageableList.getData()));
+                        return client.deleteFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                    }))
+                // File deletion
+                .assertNext(tuple -> {
+                    FileDeletionStatus deletionStatus = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
+        }));
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testCancelBatch(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getNonAzureOpenAIAsyncClient(httpClient);
+        uploadBatchFileRunner(((fileDetails, filePurpose) -> {
+            StepVerifier.create(
+                client.uploadFile(fileDetails, filePurpose)
+                    // Upload file
+                    .flatMap(uploadedFile -> {
+                        assertNotNull(uploadedFile);
+                        assertNotNull(uploadedFile.getId());
+                        return client.getFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                    })
+                    // Compare uploaded file with file from backend
+                    .flatMap(tuple -> {
+                        OpenAIFile fileFromBackend = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(uploadedFile);
+                        assertNotNull(fileFromBackend);
+                        assertFileEquals(uploadedFile, fileFromBackend);
+                        return Mono.defer(() -> {
+                            if (fileFromBackend.getStatus() == FileState.PENDING) {
+                                return Mono.delay(Duration.ofSeconds(5))
+                                    .then(client.getFile(fileFromBackend.getId())).zipWith(Mono.just(uploadedFile));
+                            } else {
+                                return Mono.just(fileFromBackend).zipWith(Mono.just(uploadedFile));
+                            }
+                        });
+                    })
+                    // Create batch with file
+                    .flatMap(tuple -> {
+                        OpenAIFile fileFromBackend = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        return client.createBatch(new BatchCreateRequest("/v1/chat/completions", fileFromBackend.getId(), "24h"))
+                            .zipWith(Mono.just(uploadedFile));
+                    })
+                    // Cancel batch
+                    .flatMap(tuple -> {
+                        Batch batch = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(batch);
+                        assertNotNull(batch.getId());
+                        return client.cancelBatch(batch.getId()).zipWith(Mono.just(uploadedFile));
+                    })
+                    // Delete uploaded file
+                    .flatMap(tuple -> {
+                        Batch cancelledBatch = tuple.getT1();
+                        OpenAIFile uploadedFile = tuple.getT2();
+                        assertNotNull(cancelledBatch);
+                        BatchStatus status = cancelledBatch.getStatus();
+                        assertTrue(status == BatchStatus.CANCELLED || status == BatchStatus.CANCELLING);
+                        return client.deleteFile(uploadedFile.getId()).zipWith(Mono.just(uploadedFile));
+                    }))
+                // File deletion
+                .assertNext(tuple -> {
+                    FileDeletionStatus deletionStatus = tuple.getT1();
+                    OpenAIFile file = tuple.getT2();
+                    assertNotNull(deletionStatus);
+                    assertNotNull(deletionStatus.getId());
+                    assertTrue(deletionStatus.isDeleted());
+                    assertEquals(file.getId(), deletionStatus.getId());
+                })
+                .verifyComplete();
         }));
     }
 }